refactor: 문의사항 관련 패키지 변경
