+++ src/main/java/com/map/gaja/temp/inquiry/InquiryCategory.java
@@ -1,4 +1,4 @@
-package com.map.gaja.temp.domain;
+package com.map.gaja.temp.inquiry;
 
 public enum InquiryCategory {
     PAYMENT, SUGGESTION, BUG, ETC
