+++ src/main/java/com/map/gaja/temp/inquiry/InquiryStatus.java
@@ -1,4 +1,4 @@
-package com.map.gaja.temp.domain;
+package com.map.gaja.temp.inquiry;
 
 public enum InquiryStatus {
     SOLVED, UNSOLVED
